SubicAds 4.0
It is a cross-platform mobile application for Subicads Advertising that
solves customer problem with their printing process, They can view
whether the status of their printing or job order is "On layout", "For
checking" or "ready to pick up". In this application the customer can
monitor their printing status through their mobile phones. An
application provides advertisements and shows live and recent videos
from the events that company covered.